% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_deltas_from_climate.R
\name{calculate_deltas_from_climate}
\alias{calculate_deltas_from_climate}
\title{calculate_deltas_from_climate}
\usage{
calculate_deltas_from_climate(
  climate_dir = NULL,
  write_dir = "outputs_calculate_delta_from_climate",
  esm_name = NULL,
  scn_name = NULL,
  crops = c("Corn", "Spring wheat", "Winter wheat", "Rice", "Soy"),
  irrigation_rainfed = c("IRR", "RFD"),
  minlat = -89.75,
  minlon = -179.75,
  monthly_growing_season = NULL,
  rollingAvgYears = 15,
  growing_season_dir = NULL,
  tas_historical = NULL,
  tas_projected = NULL,
  pr_historical = NULL,
  pr_projected = NULL,
  historical_start_year = NULL,
  projection_start_year = NULL
)
}
\arguments{
\item{climate_dir}{Default = NULL}

\item{write_dir}{Default = "outputs_calculate_delta_from_climate". Output Folder}

\item{esm_name}{Default = NULL}

\item{scn_name}{Default = NULL}

\item{crops}{Default = c("Corn", "Spring wheat", "Winter wheat", "Rice", "Soy")}

\item{irrigation_rainfed}{Default = c("IRR", "RFD")}

\item{minlat}{Default = -89.75}

\item{minlon}{Default = -179.75}

\item{monthly_growing_season}{Default = NULL. A csv file with columns latgrid, longrid, crop, irr, pmonth,    gslength, areamask}

\item{rollingAvgYears}{Default = 15}

\item{growing_season_dir}{Default = NULL}

\item{tas_historical}{= NULL. Filename of historical temperature ncdf.}

\item{tas_projected}{= NULL. Filename of projected (hot or cold) temperature ncdf.}

\item{pr_historical}{= NULL. Filename of historical precipitation flux ncdf.}

\item{pr_projected}{= NULL. Filename of projected (hot or cold) precipitation flux ncdf.}

\item{historical_start_year}{= NULL. Start year of historical data.}

\item{projection_start_year}{= NULL. Start year of projection data.}
}
\value{
number
}
\description{
Function to process climate data to calculate delta P and delta T.
}
\examples{
\dontrun{
library(osiris)
osiris::calculate_deltas_from_climate()
}
}
\keyword{test}
