% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_rainfed_yield.R
\name{cap_rainfed_yield}
\alias{cap_rainfed_yield}
\title{cap_rainfed_yield}
\usage{
cap_rainfed_yield(
  write_dir = "step5_cap_rainfed_yield",
  input_dir = NULL,
  GCAM_basin_mapping = NULL,
  GCAM_region_mapping = NULL,
  agyield_path_input = NULL,
  bioyield_path_input = NULL
)
}
\arguments{
\item{write_dir}{Default = "step5_cap_rainfed_yield". Output Folder}

\item{input_dir}{Default = NULL. Input Folder (step 4)}

\item{GCAM_basin_mapping}{Default = NULL}

\item{GCAM_region_mapping}{Default = NULL}

\item{agyield_path_input}{Default = NULL}

\item{bioyield_path_input}{Default = NULL}
}
\value{
number
}
\description{
Function that takes the defined impacts and reference AgProdChange files then
updates and saves the new AgProdChange csv and xml files.
}
\examples{
\dontrun{
library(osiris)
osiris::cap_rainfed_yield()
}
}
\keyword{test}
