% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{diagnostic_plots}
\usage{
diagnostic_plots(
  hot_ssp3_apc = NULL,
  hot_ssp5_apc = NULL,
  cold_ssp3_apc = NULL,
  cold_ssp5_apc = NULL,
  reference_apc = NULL,
  agyield = NULL,
  iso_GCAM_basin_mapping = NULL,
  write_dir = "diagnostic_plots"
)
}
\arguments{
\item{hot_ssp3_apc}{Default = NULL. Path to hot_ssp3 agprodchange csv file.}

\item{hot_ssp5_apc}{Default = NULL. Path to hot_ssp3 agprodchange csv file.}

\item{cold_ssp3_apc}{Default = NULL. Path to hot_ssp3 agprodchange csv file.}

\item{cold_ssp5_apc}{Default = NULL. Path to hot_ssp3 agprodchange csv file.}

\item{reference_apc}{Default = NULL. Path to hot_ssp3 agprodchange csv file.}

\item{agyield}{Default = NULL. Path to agyield gcamdata file L101.ag_Yield_kgm2_R_C_Y_GLU.csv.
The 2015 ag yield value is extracted from this file.}

\item{iso_GCAM_basin_mapping}{Default = NULL. Path to mapping file of GCAM basin id, GLU, ISO.}

\item{write_dir}{Default = "diagnostic_plots". Output folder for plots.}
}
\value{
number
}
\description{
Function to generate diagnostic plots to compare the agprodchange and agyield
for the different climate projections, for the specified crops and irrigation.
}
\examples{
\dontrun{
library(osiris)
osiris::diagnostic_plots()
}
}
\keyword{test}
