% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_data.R
\name{get_example_data}
\alias{get_example_data}
\title{get_example_data}
\usage{
get_example_data(
  write_dir = getwd(),
  dir_name = "Osiris_Data_Test",
  data_link = "https://zenodo.org/record/7530067/files/Osiris_Data_Test.zip?download=1"
)
}
\arguments{
\item{write_dir}{Default = getwd()}

\item{dir_name}{= "Osiris_Data_Test". Name of directory to install zip file to.}

\item{data_link}{Default = "https://zenodo.org/record/7530067/files/Osiris_Data_Test.zip?download=1"}
}
\value{
number
}
\description{
Function to download the test data files from zenodo that are needed to run Osiris.
}
\examples{
\dontrun{
library(osiris)
osiris::get_example_data()
}
}
\keyword{test}
