% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_to_basin_yield.R
\name{grid_to_basin_yield}
\alias{grid_to_basin_yield}
\title{grid_to_basin_yield}
\usage{
grid_to_basin_yield(
  carbon = NULL,
  weight_floor_ha = 1,
  fcnpath = NULL,
  inputpath = NULL,
  areapath = NULL,
  basin_grid = NULL,
  basin_id = NULL,
  outpath = getwd(),
  crops = c("maize", "rice", "soy", "wheat"),
  esm_name = "CanESM5",
  cm_name = "LPJmL",
  scn_name = "rcp8p5",
  N = 200
)
}
\arguments{
\item{carbon}{Default = NULL}

\item{weight_floor_ha}{Default = 1}

\item{fcnpath}{Default = NULL}

\item{inputpath}{Default = NULL}

\item{areapath}{Default = NULL}

\item{basin_grid}{Default = NULL}

\item{basin_id}{Default = NULL}

\item{outpath}{Default = getwd(). Output Folder}

\item{crops}{Default = c("maize", "rice", 'soy', 'wheat')}

\item{esm_name}{Default = 'CanESM5'}

\item{cm_name}{Default = 'LPJmL'}

\item{scn_name}{Default = 'rcp8p5'}

\item{N}{Default = 200. Assuming nothing is nitrogen limited and apply across grids}
}
\value{
number
}
\description{
Function that uses emulators to create gridded yield files then aggregates to
gcam basin level via MIRCA harvested areas.
}
\examples{
\dontrun{
library(osiris)
osiris::grid_to_basin_yield()
}
}
\keyword{test}
