% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{test_function}
\usage{
test_function(a, b)
}
\arguments{
\item{a}{Default = 1}

\item{b}{Default =1}
}
\value{
number
}
\description{
Description
}
\examples{
library(osiris)
osiris::test_function(1,2)
}
\keyword{test}
