% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_to_gcam_basin.R
\name{yield_to_gcam_basin}
\alias{yield_to_gcam_basin}
\title{yield_to_gcam_basin}
\usage{
yield_to_gcam_basin(
  write_dir = "outputs_yield_to_gcam_basin",
  emulated_basin_yield_dir = NULL,
  iso_GCAM_region_mapping = NULL,
  FAO_ag_mapping = NULL,
  iso_harvest_area_mapping = NULL,
  iso_GCAM_basin_mapping = NULL,
  max_CCImult = 2.5,
  min_CCImult = 0.01,
  weight_floor_ha = 1,
  rolling_avg_years = 15,
  maxHistYear = 2010,
  baseYear = 2015,
  maxFutYear = 2100
)
}
\arguments{
\item{write_dir}{Default = "outputs_yield_to_gcam_basin". Output Folder}

\item{emulated_basin_yield_dir}{Default = NULL}

\item{iso_GCAM_region_mapping}{Default = NULL}

\item{FAO_ag_mapping}{Default = NULL}

\item{iso_harvest_area_mapping}{Default = NULL}

\item{iso_GCAM_basin_mapping}{Default = NULL}

\item{max_CCImult}{Default = 2.5 Upper limit on positive climate impacts (multiplier)}

\item{min_CCImult}{Default = 0.01 Lower limit on negative climate impacts (multiplier)}

\item{weight_floor_ha}{Default = 1 Floor on area weights, in hectares. Below this climate impacts will be ignored. These are more likely than others to be problematic. 1 hectare = 0.01 km^2  = 1e-5 thou km^2, GCAM land units.}

\item{rolling_avg_years}{Default = 15 Set the number of years to define the range for rolling averages (range = 2 times this plus 1)}

\item{maxHistYear}{Default = 2010 Historical year for which to apply rolling averages}

\item{baseYear}{Default = 2015 Base year for which to apply rolling averages}

\item{maxFutYear}{Default = 2100 Future year for which to apply rolling averages}
}
\value{
number
}
\description{
Function that uses emulated yield data aggregated to GCAM basin level to create
GCAM region, basin, and irrigation level data for all GCAM commodities
}
\examples{
\dontrun{
library(osiris)
osiris::yield_to_gcam_basin()
}
}
\keyword{test}
