% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_deltas_from_climate.R
\name{calculate_deltas_from_climate}
\alias{calculate_deltas_from_climate}
\title{calculate_deltas_from_climate}
\usage{
calculate_deltas_from_climate(
  climate_dir = NULL,
  write_dir = "outputs_calculate_delta_from_climate",
  esm_name = "CanESM5",
  crops = c("Corn", "Wheat", "Rice", "Soy"),
  irrigation_rainfed = c("IRR", "RFD"),
  minlat = -87.8638,
  minlon = -179.75,
  monthly_growing_season = NULL,
  monthly_harvest_season = NULL,
  rollingAvgYears = 15,
  growing_season_dir = NULL
)
}
\arguments{
\item{climate_dir}{Default = NULL}

\item{write_dir}{Default = "outputs_calculate_delta_from_climate". Output Folder}

\item{esm_name}{Default = 'CanESM5'}

\item{crops}{Default = c("Corn", "Wheat", "Rice", "Soy")}

\item{irrigation_rainfed}{Default = c("IRR", "RFD")}

\item{minlat}{Default = -87.8638}

\item{minlon}{Default = -179.75}

\item{monthly_growing_season}{Default = NULL. A csv file with columns latgrid, longrid, crop, irr, pmonth,    gslength, areamask}

\item{monthly_harvest_season}{Default = NULL. A csv file with columns latgrid, longrid, crop, irr, pmonth,    hmonth, areamask}

\item{rollingAvgYears}{Default = 15}

\item{growing_season_dir}{Default = NULL}
}
\value{
number
}
\description{
Function to process climate data to calculate delta P and delta T.
}
\examples{
\dontrun{
library(osiris)
osiris::calculate_deltas_from_climate()
}
}
\keyword{test}
