% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_to_osiris.R
\name{wrf_to_osiris}
\alias{wrf_to_osiris}
\title{wrf_to_osiris}
\usage{
wrf_to_osiris(
  wrf_ncdf = NULL,
  osiris_ncdf = NULL,
  write_dir = "wrf_to_osiris",
  time_step = "3 hours",
  scenario = NULL
)
}
\arguments{
\item{wrf_ncdf}{Default = NULL. Path to WRF data, which should start at the beginning
of a month. You can add multiple paths, just put in a list, e.g., c("/path1", "/path2")}

\item{osiris_ncdf}{Default = NULL. Path to osiris temperature nc file.}

\item{write_dir}{Default = "wrf_to_osiris". Output Folder.}

\item{time_step}{Default = "3 hours". Other option is "1 hour".}

\item{scenario}{Default = NULL. Scenario to put in output ncdf filename.}
}
\value{
number
}
\description{
Function to extract temperature and precipitation data from WRF nc files,
reproject to match osiris CRS and resolution, calculate monthly mean temperature
and precipitation flux, and save to a new nc file.
}
\examples{
\dontrun{
library(osiris)
osiris::wrf_to_osiris()
}
}
\keyword{test}
